Move ParallelCompressedRowSparseMatrixMechanical in its own file
